feat(settings): move autocomplete model selector to Models tab
Place the autocomplete model dropdown alongside Default model and Small
model in the Models tab, as suggested by the KiloCode team.
